+++ server/src/main/java/com/hobbyt/domain/privatehome/dto/SaleCard.java
@@ -1,6 +1,6 @@
 package com.hobbyt.domain.privatehome.dto;
 
-import com.hobbyt.domain.entity.Period;
+import com.hobbyt.domain.sale.entity.Period;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
