+++ server/src/main/java/com/hobbyt/domain/sale/dto/request/SaleRequest.java
@@ -5,8 +5,8 @@ import java.util.stream.Collectors;
 
 import javax.validation.constraints.NotNull;
 
-import com.hobbyt.domain.entity.Period;
 import com.hobbyt.domain.sale.entity.Delivery;
+import com.hobbyt.domain.sale.entity.Period;
 import com.hobbyt.domain.sale.entity.Product;
 import com.hobbyt.domain.sale.entity.Sale;
 import com.hobbyt.global.entity.Account;
