+++ server/src/main/java/com/hobbyt/domain/sale/dto/response/SaleResponse.java
@@ -4,9 +4,9 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import com.hobbyt.domain.entity.Period;
 import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.domain.sale.entity.Delivery;
+import com.hobbyt.domain.sale.entity.Period;
 import com.hobbyt.domain.sale.entity.Product;
 import com.hobbyt.domain.sale.entity.Sale;
 import com.hobbyt.global.entity.Account;
