+++ server/src/main/java/com/hobbyt/domain/sale/entity/Period.java
@@ -1,9 +1,14 @@
-package com.hobbyt.domain.entity;
+package com.hobbyt.domain.sale.entity;
 
 import java.time.LocalDateTime;
 
 import javax.persistence.Embeddable;
 
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
 @Embeddable
 public class Period {
 	private LocalDateTime startedAt;
